+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.4.5] - 2019-11-27 WangXH <kurapica125@outlook.com>
+### Added
+- `GetSubTypes` Method is added to `System.Class` and `System.Interface`, will return an iterator to get all sub types directly extend or inherit the target interface or class.
+
 ## [1.4.4] - 2019-10-29 WangXH <kurapica125@outlook.com>
 ### Added
 - `IsTemporary` Property is added to System.Web.HttpSession, so you can mark a session be cleared when user closed the browser